pyproject.toml
[project]
name = "externalapi-helpers"
description = "Various ComfyUI nodes for Gemini, Replicate and OpenAI"
version = "1.0.8"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
#
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["replicate", "pillow", "numpy", "torch", "google-genai", "opencv-python", "soundfile", "openai", "av", "python-dotenv"]

[project.urls]
Repository = "https://github.com/Aryan185/ComfyUI-ExternalAPI-Helpers"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/Aryan185/ComfyUI-ExternalAPI-Helpers/wiki"
"Bug Tracker" = "https://github.com/Aryan185/ComfyUI-ExternalAPI-Helpers/issues"

[tool.comfy]
PublisherId = "aryyy18"
DisplayName = "ComfyUI-ExternalAPI-Helpers"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility
